4.20-4.22 done, 4.23 started

+++ part4/bloglist/controllers/blogposts.js
@@ -1,7 +1,8 @@
 const blogpostRouter = require('express').Router()
 const Blogpost = require('../models/blogpost')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+//const User = require('../models/user')
+//const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogpostRouter.get('/', async (request, response) => {
   /*Blogpost.find({}).then(blogposts => {
@@ -34,29 +35,29 @@ blogpostRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-const getTokenFrom = request => {
+/*const getTokenFrom = request => {
   const authorization = request.get('authorization')
   console.log(authorization)
   if (authorization && authorization.startsWith('bearer ')) {
     return authorization.replace('bearer ', '')
   }
   return null
-}
+}*/
 
-blogpostRouter.post('/', async (request, response, next) => {
+blogpostRouter.post('/', middleware.userExtractor, async (request, response, next) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  /*const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+    return response.status(401).json({ error: 'invalid token' })
+  }*/
+  //const user = await User.findById(decodedToken.id)
   //const user = await User.findById(body.userId)
   const blogpost = new Blogpost({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes | 0,
-    user: user
+    user: request.user
   })
   if(typeof blogpost.title === 'undefined' || typeof blogpost.url === 'undefined') {
     response.status(400).end()
@@ -70,24 +71,43 @@ blogpostRouter.post('/', async (request, response, next) => {
 
     try {
       const savedBlog = await blogpost.save()
-      user.blogs = user.blogs.concat(savedBlog._id)
-      await user.save()
+      request.user.blogs = request.user.blogs.concat(savedBlog._id)
+      await request.user.save()
       response.status(201).json(savedBlog)
     }
     catch(exception) {
+      //console.log(exception)
       next(exception)
     }
   }
 })
 
-blogpostRouter.delete('/:id', async (request, response, next) => {
+blogpostRouter.delete('/:id', middleware.userExtractor, async (request, response, next) => {
   /*Blogpost.findByIdAndRemove(request.params.id)
     .then(() => {
       response.status(204).end()
     })
     .catch(error => next(error))*/
+  /*const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }*/
+  //const user = await User.findById(decodedToken.id)
+  const blog = await Blogpost.findById(request.params.id)
+  try {
+    if (blog.user._id.toString() !== request.user.id) {
+      return response.status(403).json({ error: 'unauthorized delete' })
+    }
+  }
+  catch(exception) {
+    return response.status(403).json({ error: 'unauthorized delete' })
+  }
   try {
     await Blogpost.findByIdAndRemove(request.params.id)
+    //console.log(user.blogs)
+    //user.blogs = user.blogs.filter(blog => blog.id!==request.params.id)
+    request.user.blogs.pull(request.params.id)
+    await request.user.save()
     response.status(204).end()
   }
   catch(exception) {
